Split transitions component into separate components for each type of transition
