+++ sd_material_ui/_imports_.py
@@ -4,20 +4,24 @@ from .Button import Button
 from .Card import Card
 from .Checkbox import Checkbox
 from .CircularProgress import CircularProgress
+from .CollapseTransition import CollapseTransition
 from .Dialog import Dialog
 from .Divider import Divider
 from .Drawer import Drawer
 from .DropDownMenu import DropDownMenu
+from .FadeTransition import FadeTransition
 from .FontIcon import FontIcon
+from .GrowTransition import GrowTransition
 from .Paper import Paper
 from .Popover import Popover
 from .RadioButtonGroup import RadioButtonGroup
+from .SlideTransition import SlideTransition
 from .Snackbar import Snackbar
 from .Stepper import Stepper
 from .Subheader import Subheader
 from .Tabs import Tabs
 from .Toggle import Toggle
-from .Transitions import Transitions
+from .ZoomTransition import ZoomTransition
 
 __all__ = [
     "Accordion",
@@ -26,18 +30,22 @@ __all__ = [
     "Card",
     "Checkbox",
     "CircularProgress",
+    "CollapseTransition",
     "Dialog",
     "Divider",
     "Drawer",
     "DropDownMenu",
+    "FadeTransition",
     "FontIcon",
+    "GrowTransition",
     "Paper",
     "Popover",
     "RadioButtonGroup",
+    "SlideTransition",
     "Snackbar",
     "Stepper",
     "Subheader",
     "Tabs",
     "Toggle",
-    "Transitions"
+    "ZoomTransition"
 ]
