+++ src/lib/index.js
@@ -1,25 +1,25 @@
 export { default as Accordion } from './components/Accordion.react';
 // export { default as AutoComplete } from './components/AutoComplete.react';
 export { default as BottomNavigation } from './components/BottomNavigation.react';
+export { default as Button } from './components/Button.react';
 export { default as Card } from './components/Card.react';
 export { default as Checkbox } from './components/Checkbox.react';
 export { default as CircularProgress } from './components/CircularProgress.react';
+export { default as CollapseTransition } from './components/CollapseTransition.react';
 export { default as Dialog } from './components/Dialog.react';
 export { default as Divider } from './components/Divider.react';
 export { default as Drawer } from './components/Drawer.react';
 export { default as DropDownMenu } from './components/DropDownMenu.react';
-export { default as Button } from './components/Button.react';
+export { default as FadeTransition } from './components/FadeTransition.react';
 export { default as FontIcon } from './components/FontIcon.react';
+export { default as GrowTransition } from './components/GrowTransition.react';
 export { default as Paper } from './components/Paper.react';
 export { default as Popover } from './components/Popover.react';
 export { default as RadioButtonGroup } from './components/RadioButtonGroup.react';
+export { default as SlideTransition } from './components/SlideTransition.react';
 export { default as Snackbar } from './components/Snackbar.react';
 export { default as Stepper } from './components/Stepper.react';
 export { default as Subheader } from './components/Subheader.react';
 export { default as Tabs } from './components/Tabs.react';
 export { default as Toggle } from './components/Toggle.react';
-export { default as Transitions } from './components/Transitions.react';
-
-// Special components
-// export { default as Questions } from './components/Questions.react';
-// export { default as QuestionsTabs } from './components/QuestionsTabs.react';
+export { default as ZoomTransition } from './components/ZoomTransition.react';
